v2 remove most v1 IO code
