+++ packages/core/src/core.ts
@@ -2,10 +2,9 @@ import { AccessorComponentType, AccessorComponentTypeData, AccessorType, Accesso
 import { Element } from './elements/element';
 import { Material, Mesh, Node, Root } from './elements/index';
 import { Graph } from './graph/index';
-import { NodeIO, WebIO } from './io';
+import { NodeIO, WebIO } from './io/io';
 import { Logger, LoggerVerbosity } from './logger';
 import { GLTFUtil } from './util';
-import { GLTFContainer } from './v1/container';
 
 // TODO(donmccurdy): Exports are incomplete.
 export { AccessorComponentType, AccessorComponentTypeData, AccessorType, AccessorTypeData, BufferViewTarget };
@@ -15,4 +14,3 @@ export { Graph };
 export { NodeIO, WebIO };
 export { Logger, LoggerVerbosity };
 export { GLTFUtil };
-export { GLTFContainer };
